+++ crates/cosmos/cosmos-chain-components/src/encoding/mod.rs
@@ -1,5 +1,4 @@
 pub mod components;
 pub mod convert;
 pub mod encode;
-pub mod encode_mut;
 pub mod type_url;