+++ crates/cosmos/cosmos-wasm-relayer/src/context/encoding.rs
@@ -13,8 +13,8 @@ use hermes_encoding_components::traits::has_encoding::{
 use hermes_encoding_components::traits::types::encoded::HasEncodedType;
 use hermes_encoding_components::types::AsBytes;
 use hermes_protobuf_encoding_components::types::strategy::{ViaAny, ViaProtobuf};
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 use ibc::core::client::types::Height;
 use ibc_relayer_types::clients::ics07_tendermint::client_state::ClientState as TendermintClientState;
 use prost_types::Any;
@@ -103,7 +103,7 @@ mod test {
     use hermes_error::types::HermesError;
     use hermes_protobuf_encoding_components::impls::encode::buffer::EncodeProtoWithMutBuffer;
     use hermes_protobuf_encoding_components::types::strategy::ViaProtobuf;
-    use hermes_wasm_client_components::types::client_state::WasmClientState;
+    use hermes_wasm_encoding_components::types::client_state::WasmClientState;
     use ibc::core::client::types::Height;
 
     use crate::context::encoding::WasmCosmosEncoding;