+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/convert.rs
@@ -1,14 +1,15 @@
 use cgp::prelude::*;
-use hermes_cosmos_chain_components::encoding::components::CosmosEncodingComponents;
+use hermes_cosmos_chain_components::encoding::components::CosmosClientEncodingComponents;
 use hermes_cosmos_chain_components::types::tendermint::{
     ProtoTendermintClientState, ProtoTendermintConsensusState, TendermintClientState,
     TendermintConsensusState,
 };
-use hermes_wasm_client_components::impls::encoding::components::WasmEncodingComponents;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::{
-    DecodeViaWasmConsensusState, EncodeViaWasmConsensusState, WasmConsensusState,
+use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::impls::strategies::consensus_state::{
+    DecodeViaWasmConsensusState, EncodeViaWasmConsensusState,
 };
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 use ibc::core::commitment_types::merkle::MerkleProof;
 use ibc_proto::ibc::core::commitment::v1::MerkleProof as ProtoMerkleProof;
 use prost_types::Any;
@@ -31,7 +32,7 @@ delegate_components! {
             (TendermintClientState, Any),
             (Any, TendermintClientState),
         ]:
-            CosmosEncodingComponents,
+            CosmosClientEncodingComponents,
         [
             (WasmClientState, Any),
             (Any, WasmClientState),