+++ crates/cosmos/cosmos-wasm-relayer/src/encoding/type_url.rs
@@ -1,11 +1,11 @@
 use cgp::prelude::*;
-use hermes_cosmos_chain_components::encoding::components::CosmosEncodingComponents;
+use hermes_cosmos_chain_components::encoding::components::CosmosClientEncodingComponents;
 use hermes_cosmos_chain_components::types::tendermint::{
     TendermintClientState, TendermintConsensusState,
 };
-use hermes_wasm_client_components::impls::encoding::components::WasmEncodingComponents;
-use hermes_wasm_client_components::types::client_state::WasmClientState;
-use hermes_wasm_client_components::types::consensus_state::WasmConsensusState;
+use hermes_wasm_encoding_components::components::WasmEncodingComponents;
+use hermes_wasm_encoding_components::types::client_state::WasmClientState;
+use hermes_wasm_encoding_components::types::consensus_state::WasmConsensusState;
 
 pub struct WasmCosmosTypeUrlSchemas;
 
@@ -15,7 +15,7 @@ delegate_components! {
             TendermintClientState,
             TendermintConsensusState,
         ]:
-            CosmosEncodingComponents,
+            CosmosClientEncodingComponents,
         [
             WasmClientState,
             WasmConsensusState,