+++ crates/wasm/wasm-client-components/src/impls/encoding/mod.rs
@@ -1,5 +1,0 @@
-pub mod components;
-pub mod convert;
-pub mod encode_mut;
-pub mod encoder;
-pub mod type_url;