+++ crates/wasm/wasm-client-components/src/impls/mod.rs
@@ -1,2 +1,1 @@
-pub mod encoding;
 pub mod types;