+++ crates/wasm/wasm-encoding-components/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod components;
 pub mod impls;
-pub mod macros;
-pub mod traits;
 pub mod types;
-pub mod vendor;